api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
google-genai>=0.1.0
semgrep>=1.73.0
pydantic>=2.7.4
python-multipart>=0.0.9
python-multipart
